#321 DriverInput wiith listener and little changes

+++ SW/visualisation/src/src/main/java/Listeners/OnBreakSteeringWheelListener.java
@@ -1,9 +1,6 @@
 package Listeners;
 
-/**
- * Created by secured on 2016. 03. 06..
- */
 public interface OnBreakSteeringWheelListener {
     Boolean breakPushed();
-    void steeringWheelAngleChanged(float angle);
+    void steeringWheelAngleChanged(double angle);
 }
